[liba] Update the size_t type definition
size_t must be an unsigned it. Moreover, by definition, size_t is the
return type of the "sizeof" operator, which is defined by the compiler.
Both GCC and clang define a __SIZE_TYPE__ macro to let the libc header
know about the type used by the sizeof operator.

+++ poincare/src/commutative_operation.cpp
@@ -46,7 +46,7 @@ Expression * CommutativeOperation::clone() const {
 
 float CommutativeOperation::approximate(Context& context) const {
   float result = m_operands[0]->approximate(context);
-  for (size_t i=1; i<m_numberOfOperands; i++) {
+  for (int i=1; i<m_numberOfOperands; i++) {
     float next = m_operands[i]->approximate(context);
     result = this->operateApproximatevelyOn(result, next);
   }
